+++ src/app/Scenes/MyCollection/utils/filterArtistsByKeyword.ts
@@ -1,0 +1,24 @@
+import { normalizeText } from "app/utils/normalizeText"
+
+export const filterArtistsByKeyword = (
+  artists: Array<{ displayLabel: string | null }>,
+  keywordFilter: string
+) => {
+  if (keywordFilter.length < 2) {
+    return artists
+  }
+
+  const normalizedKeywordFilter = normalizeText(keywordFilter)
+
+  if (!normalizedKeywordFilter) {
+    return artists
+  }
+
+  const keywordFilterWords = normalizedKeywordFilter.split(" ")
+
+  const doAllKeywordFiltersMatch = (artist: { displayLabel: string | null }) =>
+    keywordFilterWords.filter((word) => !normalizeText(artist?.displayLabel ?? "").includes(word))
+      .length === 0
+
+  return artists.filter(doAllKeywordFiltersMatch)
+}
